Update
Rearranged point pickups in level for better game feel and player direction, fixed layout of optional room, began work on the first two rooms of level 2. adding in a few audio files but no funcitonality added for them yet sounds include (two jump sound options, a health pickup, and a noise for when the player hits the checkpoints.